transformers/appairflow9/requirements.txt
apache-airflow
azure-storage-file-share
azure-storage-blob
pandas
pyarrow
